backend/api/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiomysql==0.2.0
aioredis==2.0.1
aiosignal==1.3.2
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
billiard==4.2.1
blessed==1.22.0
boolean.py==5.0
boto3==1.40.25
CacheControl==0.14.3
cached-property==2.0.1
cachetools==6.0.0
cattrs==25.2.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
codeflash==0.17.2
codeflash-benchmark==0.2.0
colorama==0.4.6
coverage==7.10.7
crosshair-tool==0.0.96
cryptography==45.0.4
cyclonedx-python-lib==9.1.0
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
docutils==0.21.2
ecdsa==0.19.1
editor==1.6.6
execnet==2.1.1
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.6.2
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
gql==4.0.0
graphql-core==3.2.6
graphql_query==1.4.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
humanize==4.13.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inicfg==2.1.0
inquirer==3.4.1
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jinja2-pluralize==0.3.1
jinxed==1.3.0
jmespath==1.0.1
joblib==1.5.1
junitparser==4.0.2
kombu==5.5.4
lazy_loader==0.4
lia-web==0.2.3
libcst==1.8.5
librosa==0.11.0
license-expression==30.4.4
line_profiler==5.0.0
llvmlite==0.44.0
lsp-protocol==2023.0.1
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
mutagen==1.47.0
mypy_extensions==1.1.0
networkx==3.5
nicegui==2.19.0
numba==0.61.2
numpy==2.2.6
orjson==3.10.18
packaging==25.0
parametrize==0.9.0
parso==0.8.5
passlib==1.7.4
pathlib==1.0.1
Pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
posthog==6.7.6
prompt_toolkit==3.0.51
propcache==0.3.1
pscript==0.7.7
psutil==7.1.0
psutils==3.3.14
psycopg2-binary==2.9.10
puremagic==1.30
py-cpuinfo==9.0.0
py-serializable==2.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygls==1.3.1
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.5
pypdf==6.1.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-benchmark==5.1.0
pytest-mock==3.14.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
readchar==4.2.1
redis==6.2.0
regex==2025.9.18
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
ruff==0.13.1
runs==1.2.2
safetensors==0.6.2
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==5.1.0
sentinel==1.0.0
sentry-sdk==2.39.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
sqlakeyset==2.0.1746777265
SQLAlchemy==2.0.41
starlette==0.46.2
stevedore==5.5.0
strawberry-graphql==0.281.0
strawberry-sqlalchemy-mapper==0.7.0
strawberry_sqlalchemy==0.1.5
strawchemy==0.18.0
sympy==1.14.0
syrupy==4.9.1
threadpoolctl==3.6.0
timeout-decorator==0.5.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typeshed_client==2.8.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unidiff==0.7.5
unittest-xml-reporting==3.2.0
urllib3==2.5.0
uvicorn==0.34.3
vbuild==0.8.2
vine==5.1.0
wait-for-2==0.3.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Whoosh-reloaded==2.7.5
wsproto==1.2.0
xmod==1.8.1
yarl==1.20.0
z3-solver==4.15.3.0
zipp==3.23.0
pgvector==0.3.6
